VGG Version
In this ISBI challenge, we tackled the problem of automatic metastasis
detection by a novel framework with deep cascaded networks.
In the pre-processing, a statistic-info integrated threshold has been
used to filter out the Region of Interest(ROI) patches.
After that,  a VGG-16 Network has been utilized as the patch classifier.
For the post-processing, we smoothed probability map to remove the
outliers.
Finally, the metastasis was localized in the center of connected
component on the probability mask.