VGGRes Version
In this ISBI challenge, we tackled the problem of automatic metastasis
detection by a novel framework with deep cascaded networks.

This united framework consists of two components integrated in a
cascaded manner.

The first light VGG-16 network retrieves the candidates by scanning the
whole sliding image (Level 1) in a fast speed while maintaining a high
sensitivity.

The following fine-tuning deep residual neural network(ResNet-152) with
powerful residual model aims at discriminating the candidates with a
high accuracy (Level 0).

Finally, the metastasis was localized in the center of connected
component on the probability mask.